Add ConsoleApiPort to encapsulate logic for console tool
