+++ src/ApiPort/Program.cs
@@ -7,9 +7,7 @@ using Microsoft.Practices.Unity;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Globalization;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace ApiPort
 {
@@ -35,15 +33,15 @@ namespace ApiPort
 
                 try
                 {
-                    var apiPortClient = container.Resolve<ApiPortClient>();
+                    var client = container.Resolve<ConsoleApiPort>();
 
                     switch (options.Command)
                     {
                         case AppCommands.ListTargets:
-                            ListTargets(apiPortClient, container.Resolve<ITargetMapper>()).Wait();
+                            client.ListTargetsAsync().Wait();
                             break;
                         case AppCommands.AnalyzeAssemblies:
-                            AnalyzeAssembliesAsync(apiPortClient, container.Resolve<IApiPortOptions>()).Wait();
+                            client.AnalyzeAssembliesAsync().Wait();
                             break;
 #if DOCID_SEARCH
                         case AppCommands.DocIdSearch:
@@ -52,7 +50,7 @@ namespace ApiPort
                             break;
 #endif
                         case AppCommands.ListOutputFormats:
-                            ListOutputFormats(apiPortClient).Wait();
+                            client.ListOutputFormatsAsync().Wait();
                             break;
                     }
 
@@ -179,84 +177,6 @@ namespace ApiPort
             }
 
             return false;
-        }
-
-        private static async Task ListOutputFormats(ApiPortClient apiPortClient)
-        {
-            var outputFormats = await apiPortClient.GetResultFormatsAsync();
-
-            if (outputFormats.Any())
-            {
-                Console.WriteLine();
-                Console.WriteLine(LocalizedStrings.AvailableOutputFormats);
-
-                foreach (var outputFormat in outputFormats)
-                {
-                    Console.WriteLine(string.Format(LocalizedStrings.TargetsListNoVersion, outputFormat));
-                }
-            }
-        }
-
-        private static async Task ListTargets(ApiPortClient apiPortClient, ITargetMapper targetMapper)
-        {
-            const string SelectedMarker = "*";
-
-            var targets = await apiPortClient.GetTargetsAsync();
-
-            if (targets.Any())
-            {
-                Console.WriteLine();
-                Console.WriteLine(LocalizedStrings.AvailableTargets);
-
-                var expandableTargets = targets.Where(target => target.ExpandedTargets.Any());
-                var groupedTargets = targets.Where(target => !target.ExpandedTargets.Any()).GroupBy(target => target.Name);
-
-                foreach (var item in groupedTargets)
-                {
-                    Console.WriteLine(LocalizedStrings.TargetsList, item.Key, String.Join(LocalizedStrings.VersionListJoin, item.Select(v => v.Version.ToString() + (v.IsSet ? SelectedMarker : String.Empty))));
-                }
-
-                if (expandableTargets.Any())
-                {
-                    Console.WriteLine();
-                    Console.WriteLine(Microsoft.Fx.Portability.Resources.LocalizedStrings.AvailableGroupedTargets);
-
-                    foreach (var item in expandableTargets)
-                    {
-                        Console.WriteLine(LocalizedStrings.TargetsListGrouped, item.Name, String.Join(CultureInfo.CurrentCulture.TextInfo.ListSeparator + " ", item.ExpandedTargets));
-                    }
-                }
-            }
-
-            if (targetMapper.Aliases.Any())
-            {
-                Console.WriteLine();
-                Console.WriteLine(LocalizedStrings.AvailableAliases);
-
-                foreach (var alias in targetMapper.Aliases)
-                {
-                    Console.WriteLine(LocalizedStrings.TargetsListNoVersion, alias);
-                }
-            }
-
-            Console.WriteLine();
-            Console.WriteLine(LocalizedStrings.NotesOnUsage);
-            Console.WriteLine(LocalizedStrings.TargetsListNoVersion, Microsoft.Fx.Portability.Resources.LocalizedStrings.HowToSpecifyVersion);
-            Console.WriteLine();
-            Console.WriteLine(LocalizedStrings.TargetsListNoVersion, LocalizedStrings.WhatAsteriskMeans);
-        }
-
-        private static async Task AnalyzeAssembliesAsync(ApiPortClient apiPort, IApiPortOptions options)
-        {
-            var outputPaths = await apiPort.WriteAnalysisReportsAsync(options);
-
-            Console.WriteLine();
-            Console.WriteLine(LocalizedStrings.OutputWrittenTo);
-
-            foreach (var outputPath in outputPaths)
-            {
-                Console.WriteLine(outputPath);
-            }
         }
     }
 }
